Lab 6/requirements.txt
pandas==2.2.*
numpy<2
scikit-learn==1.4.*
joblib==1.*
dvc[gs]==3.*
google-cloud-bigquery==3.*
google-cloud-storage==2.*
db-dtypes==1.*
pyarrow==15.*
